Use float strokeWidth; fix bbox & coords
Change Drawable.drawForTest to accept a float strokeWidth and update all implementations (Node, Relation, Way, WayRenderer) accordingly. Way.drawForTest body is commented out for now. WayRenderer now uses a float check (!= 0.0f) before setting the stroke. Parser fixes bounding-box ordering and corrects min/max initializations and additions so bounding box values are stored in the intended order. Node.getCoordinate now returns Coordinate(lat, lon). App reCenter call argument order corrected and drawing now passes a float stroke width. These changes ensure consistent stroke typing and address bbox/coordinate ordering bugs.

+++ src/main/java/Interfaces/Drawable.java
@@ -3,5 +3,5 @@ package Interfaces;
 import java.awt.*;
 
 public interface Drawable {
-    public void drawForTest(Graphics2D gc, Color color, Integer strokeWidth);
+    public void drawForTest(Graphics2D gc, Color color, float strokeWidth);
 }
